+++ octobot_trading/exchanges/parser/positions_parser.py
@@ -123,8 +123,8 @@ class PositionsParser(Parser):
 
     def _parse_realized_pnl(self):
         self._try_to_find_and_set_decimal(
-            PositioCols.REALISED_PNL.value,
-            [ExchangeCols.REALISED_PNL.value] + RealizedPnlSynonyms.keys,
+            PositioCols.REALIZED_PNL.value,
+            [ExchangeCols.REALIZED_PNL.value] + RealizedPnlSynonyms.keys,
             use_info_sub_dict=True,
             allow_zero=True,
         )