+++ tests/personal_data/transactions/test_transaction_factory.py
@@ -62,14 +62,14 @@ async def test_create_blockchain_transaction(backtesting_trader):
     assert len(exchange_manager.exchange_personal_data.transactions_manager.transactions) == 1
 
 
-async def test_create_realised_pnl_transaction(backtesting_trader):
+async def test_create_realized_pnl_transaction(backtesting_trader):
     _, exchange_manager, _ = backtesting_trader
-    transaction = transaction_factory.create_realised_pnl_transaction(
+    transaction = transaction_factory.create_realized_pnl_transaction(
         exchange_manager,
         currency=TRANSACTION_CURRENCY,
         symbol=TRANSACTION_SYMBOL,
         side=enums.PositionSide.LONG,
-        realised_pnl=decimal.Decimal(0.1),
+        realized_pnl=decimal.Decimal(0.1),
         closed_quantity=decimal.Decimal(1),
         cumulated_closed_quantity=decimal.Decimal(2),
         first_entry_time=decimal.Decimal(3),
@@ -80,7 +80,7 @@ async def test_create_realised_pnl_transaction(backtesting_trader):
         trigger_source=enums.PNLTransactionSource.LIMIT_ORDER)
     assert transaction.currency == TRANSACTION_CURRENCY
     assert transaction.symbol == TRANSACTION_SYMBOL
-    assert transaction.realised_pnl == decimal.Decimal(0.1)
+    assert transaction.realized_pnl == decimal.Decimal(0.1)
     assert transaction.side is enums.PositionSide.LONG
     assert transaction.closed_quantity == decimal.Decimal(1)
     assert transaction.cumulated_closed_quantity == decimal.Decimal(2)